+++ Backend/Features/Basket/Commands/IAddDish.cs
@@ -11,7 +11,7 @@ namespace Backend.Features.Basket.Commands;
 /// </summary>
 /// <exception cref="ActionFailedException">When <paramref name="UserId"/>'s basket contains dishes from different to <paramref name="DishId"/> restaurant</exception>
 /// <exception cref="KeyNotFoundException"></exception>
-public sealed record AddDishCommand(Guid UserId, Guid DishId, uint Count) : IRequestWithException<EmptyResult>;
+public sealed record AddDishCommand(Guid UserId, Guid DishId, ulong Count) : IRequestWithException<EmptyResult>;
 
 [RequestHandlerInterface]
 public interface IAddDish : IRequestHandlerWithException<AddDishCommand, EmptyResult>
