+++ Backend/Features/Basket/Commands/IRemoveDish.cs
@@ -8,10 +8,10 @@ namespace Backend.Features.Basket.Commands;
 /// Decreases count of dishes in basket by <paramref name="Count"/>
 /// or removes it completely if count in basket less or equal <paramref name="Count"/>
 /// </summary>
-public sealed record RemoveDishCommand(Guid UserId, Guid DishId, uint Count = 1)
+public sealed record RemoveDishCommand(Guid UserId, Guid DishId, ulong Count = 1)
     : IRequestWithException<EmptyResult, KeyNotFoundException>
 {
-    public static RemoveDishCommand RemoveAll(Guid userId, Guid dishId) => new(userId, dishId, uint.MaxValue);
+    public static RemoveDishCommand RemoveAll(Guid userId, Guid dishId) => new(userId, dishId, ulong.MaxValue);
 }
 
 [RequestHandlerInterface]
