+++ Backend/Features/Order/Domain/DishInBasket.cs
@@ -5,5 +5,5 @@ public class DishInBasket
     public Guid UserId { get; private set; }
     public Dish Dish { get; private set; } = default!;
     public Guid DishId { get; private set; }
-    public uint Count { get; private set; }
+    public ulong Count { get; private set; }
 }
