+++ Backend/Infra/Data/Entities/DishInBasket.cs
@@ -6,5 +6,5 @@ public class DishInBasket
     public Dish Dish { get; set; } = default!;
     public Guid DishId { get; set; }
     
-    public uint Count { get; set; }
+    public ulong Count { get; set; }
 }
